update: UserService authorization

+++ siso/src/main/java/com/gabrielpdev/siso/services/UserService.java
@@ -1,8 +1,13 @@
 package com.gabrielpdev.siso.services;
 
+import java.util.Objects;
 import java.util.Optional;
 
+import com.gabrielpdev.siso.models.CustomUserDetails;
+import com.gabrielpdev.siso.models.exceptions.AuthorizationException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -17,12 +22,6 @@ import com.gabrielpdev.siso.models.exceptions.ObjectNotFoundException;
 import jakarta.transaction.Transactional;
 import jakarta.validation.Valid;
 
-/**
- * Service para entidade usuario, "API" entre controller e banco de dados
- * @author Gabriel Pinto Andrade
- * @version 1.0.0
- * @since 07/08/2024
- */
 @Service
 public class UserService {
 
@@ -32,55 +31,39 @@ public class UserService {
     @Autowired
     BCryptPasswordEncoder bCryptPasswordEncoder;
 
-    /**
-     * Procura um usuario por id
-     * @param id id do usuario
-     * @return Usuario
-     */
     public User findById(Long id) {
+        if(!(isSelf(id) || isAdmin())) throw new AuthorizationException("Acesso negado.");
         Optional<User> obj = this.userRepository.findById(id);
         return obj.orElseThrow(() -> new ObjectNotFoundException("Usuário {id:"+id+"} não encontrado"));
     }
 
-    public User findByUsername(String username) {
-        Optional<User> obj = this.userRepository.findByUsername(username);
-        return obj.orElseThrow(() -> new UsernameNotFoundException("Usuário não encontrado"));
-    }
-
-    /**
-     * Cria um usuario
-     * @param obj Objeto Usuario
-     */
     @Transactional
     public void createUser(User obj) {
+        if(!isAdmin()) throw new AuthorizationException("Acesso negado.");
         obj.setId(null);
         obj.setPassword(bCryptPasswordEncoder.encode(obj.getPassword()));
-        obj.addProfile("USER");
         obj.setAtivo(true);
         this.userRepository.save(obj);
     }
 
-    /**
-     * Atualiza um usuario
-     * @param obj Objeto Usuario
-     */
     @Transactional
     public void updateUser(User obj) {
+        if(hasRoot(obj) ||  !(isSelf(obj.getId()) || isAdmin())) throw new AuthorizationException("Acesso negado.");
         User newObj = this.findById(obj.getId());
-        newObj.setPassword(bCryptPasswordEncoder.encode(obj.getPassword()));
+        if(isSelf(obj.getId())) {
+            newObj.setPassword(bCryptPasswordEncoder.encode(obj.getPassword()));
+        }
         newObj.setEmail(obj.getEmail());
-        newObj.setProfiles(obj.getProfiles());
-        newObj.setAtivo(obj.getAtivo());
+        if(isAdmin()) {
+            newObj.setProfiles(obj.getProfiles());
+            newObj.setAtivo(obj.getAtivo());
+        }
         this.userRepository.save(newObj);
     }
 
-    /**
-     * Deleta um usuario do banco por id
-     * @param id id do usuario a ser deletado
-     */
     public void deleteUserById(Long id) {
         findById(id);
-
+        if(!isAdmin()) throw new AuthorizationException("Acesso negado.");
         try {
             this.userRepository.deleteById(id);
         } catch (Exception e) {
@@ -104,4 +87,36 @@ public class UserService {
         user.setAtivo(obj.getAtivo());
         return user;
     }
+
+    public CustomUserDetails authenticated() {
+        try {
+            return (CustomUserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        } catch (Exception e) {
+            return null;
+        }
+    }
+
+    public boolean isAuthenticated() {
+        if(Objects.isNull(authenticated())) {
+            throw new AuthorizationException("Acesso negado.");
+        }
+        return true;
+    }
+
+    public boolean isSelf(Long id) {
+        return id.equals(authenticated().getId()) && isAuthenticated();
+    }
+
+    public boolean isAdmin() {
+        return authenticated().hasRole("ADMIN") && isAuthenticated();
+    }
+
+    public boolean isRoot() {
+        return authenticated().hasRole("ROOT") && isAuthenticated();
+    }
+
+    public boolean hasRoot(User obj) {
+        return obj.getProfiles().contains("ROOT");
+    }
+
 }
